refactor: move EducationPreview component to its own file
